Move storage.stream._StreamToken to types.RoomStreamToken
